% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterator}

\item{x}{sample length}

\item{fun}{the function you're using}

\item{alpha}{confidence interval bound}

\item{cx}{text size}

\item{...}{allows for extra arguments to be passed to the hist function.}
}
\value{
a list with xstat contained in it along with all other existing quantities
}
\description{
Bootstrap function
}
\examples{
myboot2(x=c(1,1,1,2,2,2,2,3,3,3,4,4), col="Green")
}
