% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Histogram of density sample mean}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{the iterator}

\item{a}{lower bound}

\item{b}{upper bound}
}
\value{
a graph that is a histogram
}
\description{
Histogram of density sample mean
}
\examples{
mycltu(100, 1000, 0, 10)
}
