% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{likelihood function that you want to maximize. It takes x and param as inputs.}

\item{x}{A vector of data points or observations.
These are the values over which the likelihood function is evaluated.}

\item{param}{A vector of parameter values that the function will test
to find the value that maximizes the likelihood.}

\item{...}{Additional optional arguments passed to the plot function.}
}
\value{
a list of certain parameters
}
\description{
mymaxlik
}
\examples{
logbin=function(x,param) log(dbinom(x,prob=param,size=50))
mymaxlik(logbin,x=c(1,2,9,7,7,6,9,8),param=seq(0,1,length=1000),main="Maximum Likelihood, n=20")
}
