% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{upper limit}
}
\value{
the calculated mean, sd and area in a list
}
\description{
myncurve
}
\examples{
myncurve(mu = 10, sigma = 5, a = 6)
}
