% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Title p-value function}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{calculated t-statistic}

\item{xmax}{maximun x value for plotting}

\item{n}{sample size}

\item{alpha}{significance level}
}
\value{
list containing q and the p-value
}
\description{
Title p-value function
}
\examples{
mypvalue(2.1, xmax = 4, n = 30, alpha = 0.05)
}
