% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{the probability that a plane will be truly overbooked}

\item{p}{probability that someone that got a ticket will show up}
}
\value{
the number of tickets
}
\description{
ntickets
}
\examples{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
