% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygen.R
\name{poly}
\alias{poly}
\title{polynomial generator}
\usage{
poly(degree, coefficients, x)
}
\arguments{
\item{degree}{degree of polynomial}

\item{coefficients}{your given coefficients}

\item{x}{the variable}
}
\value{
a polynomial
}
\description{
polynomial generator
}
\examples{
poly(5, (1, 2, 3, 4, 5, 6), x)
}
