% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{the iterator}

\item{n}{number of times it is tested}

\item{p}{the probability}
}
\value{
a table of relative frequencies
}
\description{
mybin function
}
\examples{
mybin(iter=10000,n=100, p=0.5)
}
